Moving HUD MBs to right folder/namespace
