+++ NitroxClient/Communication/Packets/Processors/ServerStoppedProcessor.cs
@@ -1,6 +1,6 @@
 ﻿using NitroxClient.Communication.Abstract;
 using NitroxClient.Communication.Packets.Processors.Abstract;
-using NitroxClient.MonoBehaviours.Gui.InGame;
+using NitroxClient.MonoBehaviours.Gui.Modals;
 using NitroxModel.Packets;
 
 namespace NitroxClient.Communication.Packets.Processors;
