+++ NitroxClient/GameLogic/HUD/PdaTabs/uGUI_PlayerPingEntry.cs
@@ -2,18 +2,18 @@ using System.Collections;
 using System.Collections.Generic;
 using NitroxClient.Communication.Abstract;
 using NitroxClient.Communication.Packets.Processors;
-using NitroxClient.GameLogic.HUD.PdaTabs;
+using NitroxClient.GameLogic.HUD.Components;
 using NitroxClient.GameLogic.PlayerLogic.PlayerModel.Abstract;
-using NitroxClient.MonoBehaviours.Gui.InGame;
+using NitroxClient.MonoBehaviours.Gui.Modals;
+using NitroxModel_Subnautica.DataStructures;
 using NitroxModel.Core;
 using NitroxModel.DataStructures.GameLogic;
 using NitroxModel.Packets;
-using NitroxModel_Subnautica.DataStructures;
 using UnityEngine;
 using UnityEngine.UI;
 using UWE;
 
-namespace NitroxClient.GameLogic.HUD.Components;
+namespace NitroxClient.GameLogic.HUD.PdaTabs;
 
 public class uGUI_PlayerPingEntry : uGUI_PingEntry
 {
@@ -195,7 +195,7 @@ public class uGUI_PlayerPingEntry : uGUI_PingEntry
     private IEnumerator AssignSprites()
     {
         yield return new WaitUntil(() => parent.FinishedLoadingAssets);
-        
+
         // NB: Those textures MUST be exported with a Texture Type of "Sprite (2D and UI)", else they will look blurry not matter what
         // NB 2: Those textures for the buttons are scaled 68x61 but the image inside but not hit the borders to have a better render
         MutedSprite = parent.GetSprite("muted@3x");
@@ -220,7 +220,7 @@ public class uGUI_PlayerPingEntry : uGUI_PingEntry
     private void RefreshButtonsVisibility()
     {
         LocalPlayer localPlayer = NitroxServiceLocator.LocateService<LocalPlayer>();
-        
+
         bool isNotLocalPlayer = !IsLocalPlayer;
         // We don't want any control buttons to appear for the local player
         ShowObject.SetActive(isNotLocalPlayer);
