+++ NitroxClient/MonoBehaviours/Gui/Modals/ConfirmModal.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace NitroxClient.MonoBehaviours.Gui.InGame;
+namespace NitroxClient.MonoBehaviours.Gui.Modals;
 
 public class ConfirmModal : Modal
 {
