+++ NitroxPatcher/Patches/Dynamic/Bench_OnHandClick_Patch.cs
@@ -1,8 +1,8 @@
 ﻿using System.Reflection;
 using NitroxClient.GameLogic;
-using NitroxClient.GameLogic.HUD.Components;
 using NitroxClient.GameLogic.Simulation;
 using NitroxClient.MonoBehaviours;
+using NitroxClient.MonoBehaviours.Gui.HUD;
 using NitroxModel.DataStructures;
 using NitroxModel.Helper;
